Sorry another breaking change hopefully the last for a while. I did a stupid design before that I had to change. Allow items to define tool actions. Fix armor not being able to use custom materials.

+++ src/main/java/dev/gigaherz/jsonthings/util/parse/value/Any.java
@@ -17,7 +17,7 @@ public interface Any
 
     IntValue intValue();
 
-    IntValue longValue();
+    LongValue longValue();
 
     FloatValue floatValue();
